requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
spacy
python-multipart
pypdf
langchain-community
langchain-huggingface
